refactor: create folder for database related files
