+++ apps/web/src/data/database/index.ts
@@ -1,4 +1,4 @@
-import * as schema from "./shared/schema";
+import * as schema from "../shared/schema";
 import "@/lib/envConfig";
 import { drizzle } from "drizzle-orm/libsql";
