+++ apps/web/src/data/domains/firetrucks/addToIncident.ts
@@ -1,5 +1,5 @@
 import { eq } from "drizzle-orm";
-import db from "../../";
+import db from "../../database";
 import { firetrucks } from "../../shared/schema";
 
 const addToIncident = async (firetruckId: string, incidentId: string) => {
